feat: map in assistant details page
