+++ src/modules/map/index.tsx
@@ -1,3 +1,6 @@
 export { Map } from './map';
 export { MapWithHeader } from './map-with-header';
 export { MapHeader } from './map-header';
+
+export * from './utils';
+export * from './types';
