+++ src/modules/map/use-map-legs.ts
@@ -1,10 +1,10 @@
 import { ContrastColor, Theme } from '@atb-as/theme';
 import { transportModeToColor } from '@atb/modules/transport-mode';
-import { MapLegType } from '@atb/page-modules/departures';
 import { useCallback, useEffect } from 'react';
 import hexToRgba from 'hex-to-rgba';
 import { useTheme } from '@atb/modules/theme';
 import { ComponentText, useTranslation } from '@atb/translations';
+import { MapLegType } from './types';
 
 export const useMapLegs = (
   mapRef: React.MutableRefObject<mapboxgl.Map | undefined>,
