+++ src/page-modules/assistant/details/index.tsx
@@ -5,13 +5,18 @@ import TripSection from './trip-section';
 import style from './details.module.css';
 import DetailsHeader from './details-header';
 import { ButtonLink } from '@atb/components/button';
+import { Map } from '@atb/modules/map';
+import { secondsToDuration } from '@atb/utils/date';
+import { Typo } from '@atb/components/typography';
 
 export type AssistantDetailsProps = {
   tripPattern: TripPatternWithDetails;
 };
 
 export function AssistantDetails({ tripPattern }: AssistantDetailsProps) {
-  const { t } = useTranslation();
+  const { t, language } = useTranslation();
+
+  const mapLegs = tripPattern.legs.map((leg) => leg.mapLegs).flat();
   return (
     <div className={style.container}>
       <div className={style.headerContainer}>
@@ -26,6 +31,38 @@ export function AssistantDetails({ tripPattern }: AssistantDetailsProps) {
           icon={{ left: <MonoIcon icon="navigation/ArrowLeft" /> }}
         />
         <DetailsHeader tripPattern={tripPattern} />
+      </div>
+      <div className={style.mapContainer}>
+        <Map
+          position={{
+            lon: tripPattern.legs[0].fromPlace.longitude,
+            lat: tripPattern.legs[0].fromPlace.latitude,
+          }}
+          initialZoom={13.5}
+          mapLegs={mapLegs}
+        />
+        <div className={style.tripDetails}>
+          <div className={style.duration}>
+            <MonoIcon icon="time/Duration" />
+            <Typo.p textType="body__primary">
+              {t(
+                PageText.Assistant.details.mapSection.travelTime(
+                  secondsToDuration(tripPattern.duration, language),
+                ),
+              )}
+            </Typo.p>
+          </div>
+          <div className={style.walkDistance}>
+            <MonoIcon icon="transportation/Walk" />
+            <Typo.p textType="body__primary">
+              {t(
+                PageText.Assistant.details.mapSection.walkDistance(
+                  String(tripPattern.walkDistance),
+                ),
+              )}
+            </Typo.p>
+          </div>
+        </div>
       </div>
       <div className={style.tripContainer}>
         {tripPattern.legs.map((leg, index) => (
