+++ src/page-modules/departures/server/journey-planner/index.ts
@@ -33,18 +33,13 @@ import {
   Departure,
   ServiceJourneyData,
   serviceJourneySchema,
-  MapLegType,
 } from './validators';
-import {
-  TransportMode as GraphQlTransportMode,
-  PointsOnLink as GraphQlPointsOnLink,
-} from '@atb/modules/graphql-types';
+import { TransportMode as GraphQlTransportMode } from '@atb/modules/graphql-types';
 import { PtSituationElement as GraphQlSituation } from '@atb/modules/graphql-types';
 import {
   type TransportModeType,
   isTransportModeType,
   filterGraphQlTransportModes,
-  type TransportSubmodeType,
 } from '@atb/modules/transport-mode';
 import {
   ServiceJourneyWithEstimatedCallsDocument,
@@ -52,9 +47,8 @@ import {
   ServiceJourneyWithEstimatedCallsQueryVariables,
 } from './journey-gql/service-journey-with-estimated-calls.generated';
 import { formatISO } from 'date-fns';
-import polyline from '@mapbox/polyline';
-import haversineDistance from 'haversine-distance';
 import { Situation } from '@atb/modules/situations';
+import { mapToMapLegs } from '@atb/modules/map';
 
 export type DepartureInput = {
   id: string;
@@ -85,12 +79,7 @@ export type ServiceJourneyInput = {
   fromQuayId: string;
 };
 
-export type {
-  StopPlaceInfo,
-  NearestStopPlacesData,
-  StopPlaceWithDistance,
-  MapLegType,
-};
+export type { StopPlaceInfo, NearestStopPlacesData, StopPlaceWithDistance };
 
 export type JourneyPlannerApi = {
   departures(input: DepartureInput): Promise<DepartureData>;
@@ -428,57 +417,6 @@ type RecursivePartial<T> = {
     : T[P] extends object | undefined
     ? RecursivePartial<T[P]>
     : T[P];
-};
-
-const mapToMapLegs = (
-  pointsOnLink: GraphQlPointsOnLink | undefined,
-  transportMode: TransportModeType,
-  transportSubmode: TransportSubmodeType | undefined,
-  fromStopPlace?: { id: string; latitude?: number; longitude?: number },
-) => {
-  if (!pointsOnLink || !pointsOnLink.points) return [];
-  const points = polyline.decode(pointsOnLink.points);
-  const fromCoordinates: [number, number] = [
-    fromStopPlace?.latitude ?? 0,
-    fromStopPlace?.longitude ?? 0,
-  ];
-
-  const mainStartIndex = findIndex(points, fromCoordinates);
-  const mainEndIndex = points.length - 1;
-
-  const beforeLegCoords = points.slice(0, mainStartIndex + 1);
-  const mainLegCoords = points.slice(mainStartIndex, mainEndIndex + 1);
-  const afterLegCoords = points.slice(mainEndIndex);
-
-  const toMapLeg = (item: [number, number][], faded: boolean) => {
-    return {
-      transportMode,
-      transportSubmode,
-      faded,
-      points: item,
-    };
-  };
-
-  const mapLegs: MapLegType[] = [
-    toMapLeg(beforeLegCoords, true),
-    toMapLeg(mainLegCoords, false),
-    toMapLeg(afterLegCoords, true),
-  ];
-
-  return mapLegs;
-};
-
-const findIndex = (
-  array: [number, number][],
-  quayCoords: [number, number],
-): number => {
-  return array.reduce(
-    (closest, t, index) => {
-      const distance = haversineDistance(t, quayCoords);
-      return distance < closest.distance ? { index, distance } : closest;
-    },
-    { index: -1, distance: 100 },
-  ).index;
 };
 
 const mapAndFilterDuplicateGraphQlSituations = (
